[quant] [quant] AO migration of the `quant_types.py` (phase 1)
AO Team is migrating the existing torch.quantization into torch.ao.quantization. We are doing it one file at a time to make sure that the internal callsites are updated properly.
This migrates the quant_type.py from torch.quantization to torch.ao.quantization.
At this point both locations will be supported. Eventually the torch.quantization will be deprecated.

Differential Revision: [D30898422](https://our.internmc.facebook.com/intern/diff/D30898422/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D30898422/)!

[ghstack-poisoned]
