+++ torch/ao/quantization/__init__.py
@@ -25,3 +25,8 @@ from .fuse_modules import (
     fuse_modules,
     get_fuser_method,
 )  # noqa: F401
+
+from .quant_type import (
+    QuantType,
+    quant_type_to_str,
+)
