+++ torch/quantization/quant_type.py
@@ -1,19 +1,11 @@
-import enum
+# flake8: noqa: F401
+r"""
+This file is in the process of migration to `torch/ao/quantization`, and
+is kept here for compatibility while the migration process is ongoing.
+If you are adding a new entry/functionality, please, add it to the
+`torch/ao/quantization/quant_type.py`, while adding an import statement
+here.
+"""
 
-# Quantization type (dynamic quantization, static quantization).
-# Should match the c++ enum in quantization_type.h
-class QuantType(enum.IntEnum):
-    DYNAMIC = 0
-    STATIC = 1
-    QAT = 2
-    WEIGHT_ONLY = 3
-
-
-def quant_type_to_str(quant_type):
-    m = {
-        QuantType.STATIC: "static",
-        QuantType.DYNAMIC: "dynamic",
-        QuantType.QAT: "qat",
-        QuantType.WEIGHT_ONLY: "weight_only",
-    }
-    return m[quant_type]
+from torch.ao.quantization.quant_type import QuantType
+from torch.ao.quantization.quant_type import quant_type_to_str
