+++ torch/quantization/utils.py
@@ -4,7 +4,7 @@ Utils shared by different modes of quantization (eager/graph)
 import warnings
 import functools
 import torch
-from .quant_type import QuantType, quant_type_to_str
+from torch.ao.quantization.quant_type import QuantType, quant_type_to_str
 from typing import Tuple, Any
 
 def get_combined_dict(default_dict, additional_dict):
